feat:完成日志模块

+++ lib-common/src/main/java/com/univolunteer/common/domain/entity/Users.java
@@ -13,7 +13,7 @@ public class Users {
     private String email;               // 邮箱地址
     private String phone;               // 手机号
     private UserRoleEnum role;               // 用户主角色（0志愿者，1招募方，2管理员）
-    private Long status;             // 账号状态（1正常，0冻结）
+    private Integer status;             // 账号状态（1正常，0冻结）
     private LocalDateTime lastLoginTime; // 最后登录时间
     private LocalDateTime createTime;    // 创建时间
     private Long organizationId;  //组织id
